% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSAsim-internals.R
\name{getVarFromFormula}
\alias{getVarFromFormula}
\alias{iChk4Namespace}
\alias{iCheckRStudio}
\title{Internal functions used in FSAsim.}
\usage{
getVarFromFormula(formula, data, expNumVars = NULL)
}
\description{
Internal functions used in FSAsim
}
\keyword{internal}
