% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{FSAsim}
\alias{FSAsim}
\alias{FSAsim-package}
\title{Fisheries stock assessment simulation functions.}
\description{
Functions to simulate data for fisheries stock assessment methods.
}
\details{
This package can be used to simulate data for fisheries stock assessment methods.
}

