% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srSim.R
\name{srSim}
\alias{srSim}
\title{Dynamic plots to explore typical fisheries stock-recruitment models.}
\usage{
srSim(type = c("BevertonHolt", "Ricker"), param = 1, max.S = 500,
  max.R = 1000)
}
\arguments{
\item{type}{A string that indicates the type of the stock-recruitment model. Must be one of \code{"BevertonHolt"}, \code{"Ricker"}, \code{"Shepherd"}, or \code{"SailaLorda"}.}

\item{param}{A numeric that indicates the parameterization of the stock-recruitment model type.}

\item{max.S}{A single numeric that indicates the maximum spawning stock to use for scaling the x-axis. Ignored if \code{S} is not NULL.}

\item{max.R}{A single numeric that indicates the maximum recruitment to use for scaling the y-axis. Ignored if \code{S} is not NULL.}
}
\value{
None. However a dynamic graphic connected to slider bar controls of the \bold{\sQuote{a}}, \bold{\sQuote{b}}, or \bold{\sQuote{Rp}} parameters specific to the chosen stock-recruit model.
}
\description{
Plots modeled number of recruits versus stock size for common parameterizations of the Beverton-Holt and Rickerstock-recruit models. Slider bars are used to dynamically alter the parameters of each model.
}
\details{
This function can be used to explore the dynamics of stock-recruitment models for various parameter choices. In these instances of model exploration the \code{S=} and \code{R=} arguments should be (left) set at \code{NULL}.

The \code{type=} argument is used to choose either the \code{"BevertonHolt"} or \code{"Ricker"} stock-recruitment models. Common parameterizations of the \code{"BevertonHolt"} and \code{"Ricker"} models can be chosen with \code{param=}. Four paramaterizations of the Beverton-Holt model and three parameterizations of the Ricker model are allowed. See \code{srFunShow} described in \code{\link[FSA]{stockRecruitment}} to see equations for each model.
}
\examples{
## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {
# Beverton-Holt models
srSim()
srSim(param=2)
srSim(param=3)
srSim(param=4)

# Ricker models
srSim(type="Ricker")
srSim(type="Ricker",param=2)
srSim(type="Ricker",param=3)
}  ## END INTERACTIVE MODE

}
\seealso{
\code{srStartsDP} and \code{srFunShow} described in \code{\link[FSA]{stockRecruitment}}.
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{iplot}
