% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbStartsDP.R
\name{vbStartsDP}
\alias{vbStartsDP}
\title{Uses dynamic plots to find reasonable starting values for a von Bertalanffy growth function.}
\usage{
vbStartsDP(formula, data = NULL, type = c("Typical", "typical",
  "Traditional", "traditional", "BevertonHolt", "Original", "original",
  "vonBertalanffy", "GQ", "GallucciQuinn", "Mooij", "Weisberg", "Schnute",
  "Francis", "Somers", "Somers2"), ages2use = NULL, methEV = c("means",
  "poly"), meth0 = c("yngAge", "poly"), ...)
}
\arguments{
\item{formula}{A formula of the form \code{len~age}.}

\item{data}{A data frame that contains the variables in \code{formula}.}

\item{type}{A string that indicates the parameterization of the von Bertalanffy model.}

\item{ages2use}{A numerical vector of the two ages to be used in the Schnute or Francis paramaterizations. See \code{\link[FSA]{vbStarts}}.}

\item{methEV}{A string that indicates how the lengths of the two ages in the Schnute paramaterization or the three ages in the Francis paramaterization should be derived. See \code{\link[FSA]{vbStarts}}.}

\item{meth0}{A string that indicates how the t0 and L0 paramaters should be derived. See \code{\link[FSA]{vbStarts}}.}

\item{\dots}{Further arguments passed to the methods.}
}
\value{
None, but a dynamic plot is constructed.
}
\description{
Uses dynamic plots to find reasonable starting values for the parameters in a specific parameterization of the von Bertalanffy growth function.
}
\details{
Starting values can be obtained by plotting the data with the model superimposed but tied to slider bars that allow the parameters to be interactively changed. One can change the parameters until a reasonable fit is observed and then use those valeus as starting values. The initial parameters for the slider bars are the starting values constructed with \code{\link[FSA]{vbStarts}}. It should be noted that the dynamic plot may show an error of \dQuote{[tcl] can't get device image}, but the plot will correctly update if the slider bar is adjusted.
}
\note{
The \sQuote{original} and \sQuote{vonBertalanffy} and the \sQuote{typical} and \sQuote{BevertonHolt} parameterizations are synonymous.
}
\section{IFAR Chapter}{
 12-Individual Growth.
}

\examples{
## Dynamic Plots Method -- ONLY RUN IN INTERACTIVE MODE
if (interactive()) {
  require(FSA)
  data(SpotVA1)
  vbStartsDP(tl~age,data=SpotVA1)
  vbStartsDP(tl~age,data=SpotVA1,type="typical")
  vbStartsDP(tl~age,data=SpotVA1,type="Typical")
  vbStartsDP(tl~age,data=SpotVA1,type="traditional")
  vbStartsDP(tl~age,data=SpotVA1,type="Traditional")
  vbStartsDP(tl~age,data=SpotVA1,type="BevertonHolt")
  vbStartsDP(tl~age,data=SpotVA1,type="original")
  vbStartsDP(tl~age,data=SpotVA1,type="Original")
  vbStartsDP(tl~age,data=SpotVA1,type="vonBertalanffy")
  vbStartsDP(tl~age,data=SpotVA1,type="GQ")
  vbStartsDP(tl~age,data=SpotVA1,type="GallucciQuinn")
  vbStartsDP(tl~age,data=SpotVA1,type="Mooij")
  vbStartsDP(tl~age,data=SpotVA1,type="Weisberg")
  vbStartsDP(tl~age,data=SpotVA1,type="Francis",ages2use=c(0,5))
  vbStartsDP(tl~age,data=SpotVA1,type="Schnute",ages2use=c(0,5))
} 

}
\references{
Ogle, D.H. 2016. \href{http://derekogle.com/IFAR}{Introductory Fisheries Analyses with R}. Chapman & Hall/CRC, Boca Raton, FL.

See references in \code{\link{vbFuns}}.
}
\seealso{
See \code{\link[FSA]{vbStarts}} for related functionality.
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
\keyword{manip}
