% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSAsim.R
\docType{package}
\name{FSAsim}
\alias{FSAsim}
\alias{FSAsim-package}
\title{Fisheries stock assessment simulation functions.}
\description{
Functions to simulate data for fisheries stock assessment methods.
}
\details{
This package can be used to simulate data for fisheries stock assessment methods.
}

